// src/app/shared/models/vehicle/vehicle-param.model.ts
export class VehicleParamModel {
  createdBy: string;
  createdOn: string;
  driverId: string;
  modifiedBy: string;
  modifiedOn: string;
  transporterId: string;
  vehicleId: string;
  vehicleNumber: string;
  vehicleRCLink: string;
  vehicleStatus: string;
  vehicleType: string;
  driverFirstName: string;
  driverLastName: string;
  transporterName: string;
}